services/auth/pyproject.toml
[project]
name = "mint-auth"
version = "2.0.0"
description = "auth service for mint"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "argon2-cffi>=25.1.0",
    "fastapi[all]>=0.135.1",
    "joserfc>=1.6.3",
    "pydantic-settings==2.12.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.48",
    "sreekarnv-fastauth>=0.5.1",
]

[project.scripts]
dev = "auth.main:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/auth"]
